refactor: reorganize Firebase setup and move to utils; remove unused firebase.ts
