+++ app/utils/firebase.ts
@@ -3,6 +3,7 @@ import { initializeApp } from "firebase/app";
 import { getFirestore } from "firebase/firestore";
 import { getPerformance } from "firebase/performance";
 import { getAnalytics } from "firebase/analytics";
+
 // TODO: Add SDKs for Firebase products that you want to use
 // https://firebase.google.com/docs/web/setup#available-libraries
 
@@ -18,7 +19,15 @@ const firebaseConfig = {
 
 const app = initializeApp(firebaseConfig);
 const db = getFirestore(app);
-const perf = getPerformance(app);
-const analytics = getAnalytics(app);
 
-export { db , perf, analytics};
+try {
+  if (typeof window !== 'undefined') {
+    getPerformance(app);
+    getAnalytics(app);
+  }
+} catch (error) {
+  console.error('Firebase Performance initialization error:', error);
+}
+
+
+export { db };
